# editor/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "editor API",
    "version": "0.1.0",
    "description": "Jinja2 テンプレート GUI 編集アプリのバックエンド API。\n\n`@editor/shared` の Repository 契約(Template / Auth / User / History / Part)を REST として設計したもの。一部は Phase 2 の設計(ハンドラ未実装)。"
  },
  "servers": [
    {
      "url": "/api",
      "description": "同一オリジン (server が SPA も配信)"
    }
  ],
...
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/parts/classification-options": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "パーツ分類のカスケード候補",
        "operationId": "getPartClassificationOptions",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "majorClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "middleClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minorClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "editionType",
            "schema": {
              "description": "テンプレートの版種。指定時は対象版種が空か一致するパーツだけ",
              "type": "string"
            },
            "description": "テンプレートの版種。指定時は対象版種が空か一致するパーツだけ"
          }
        ],
        "responses": {
          "200": {
            "description": "各段階の候補",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartClassificationOptions"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/parts": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "パーツ一覧(分類で絞り込み)",
        "operationId": "listParts",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "majorClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "middleClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minorClass",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "editionType",
            "schema": {
              "description": "テンプレートの版種。指定時は対象版種が空か一致するパーツだけ",
              "type": "string"
            },
            "description": "テンプレートの版種。指定時は対象版種が空か一致するパーツだけ"
          }
        ],
        "responses": {
          "200": {
            "description": "パーツの配列",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartCatalogItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
    "/templates/{templateId}/part-history": {
      "get": {
        "tags": [
          "parts"
        ],
        "summary": "パーツ単位の編集履歴(版インスタンス単位の全件)",
        "operationId": "listPartHistory",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "パーツ履歴の配列",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartHistoryEntry"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "404": {
            "description": "対象が存在しない (kind=not_found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "parts"
        ],
        "summary": "パーツ単位の編集を記録",
        "operationId": "recordPartChange",
        "parameters": [
          {
            "in": "path",
            "name": "templateId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPartChangeRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "記録完了"
          },
          "400": {
            "description": "リクエスト検証エラー (kind=validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          },
          "401": {
            "description": "未認証 (kind=unauthorized)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecordPartChangeRequest": {
        "type": "object",
        "properties": {
          "partKey": {
            "type": "string",
            "minLength": 1,
            "description": "パーツ構造キー(文書全体でのアンカー#通し番号)"
          },
          "change": {
            "type": "string"
          },
          "id": {
            "description": "履歴 1 件の id(web が採番。重複表示の除去に使う)",
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "partKey",
          "change"
        ]
      },
...
      "AppError": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AppErrorKind"
          },
          "message": {
            "type": "string",
            "description": "ユーザ向け(JP)。常に表示して安全"
          },
          "code": {
            "description": "機械可読コード 例: 'USER_DISABLED'",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "message"
        ],
        "additionalProperties": false
      },
      "AppErrorKind": {
        "type": "string",
        "enum": [
          "not_found",
          "validation",
          "unauthorized",
          "forbidden",
          "conflict",
          "network",
          "unexpected"
        ]
      },
...
      "PartClassificationOptions": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "majorClasses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "middleClasses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minorClasses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "categories",
          "majorClasses",
          "middleClasses",
          "minorClasses"
        ],
        "additionalProperties": false
      },
      "PartCatalogItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "安定したパーツ ID(SQL 主キー相当)。挿入したコンポーネントの data-part-id にも使う"
          },
          "classification": {
            "$ref": "#/components/schemas/PartClassification"
          },
          "name": {
            "type": "string",
            "description": "名称(利用者向け)"
          },
          "description": {
            "type": "string",
            "description": "説明(利用者向け)"
          },
          "usageNotes": {
            "type": "string",
            "description": "使用上の注意"
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "type": "string",
            "description": "キャンバスに挿入する GrapesJS 用 HTML 断片"
          },
          "syncDefault": {
            "description": "交付版⇄全体版 自動同期の既定。null/欠落 = 未判断(同期しない)",
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartSyncDefault"
              },
              {
                "type": "null"
              }
            ]
          },
          "masterReflectDefault": {
            "description": "注記マスタ書き戻しの既定。null/欠落 = 未判断(反映しない)",
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartMasterReflectDefault"
              },
              {
                "type": "null"
              }
            ]
          },
          "targetEdition": {
            "description": "対象の版種。null/欠落 = 両版共通。一覧の絞り込み表示だけに使い、同期は syncDefault で決める",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "classification",
          "name",
          "description",
          "usageNotes",
          "updatedAt",
          "updatedBy",
          "content"
        ],
        "additionalProperties": false
      },
      "PartClassification": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "description": "カテゴリ(最上位)"
          },
          "majorClass": {
            "type": "string",
            "description": "大分類"
          },
          "middleClass": {
            "type": "string",
            "description": "中分類"
          },
          "minorClass": {
            "type": "string",
            "description": "小分類"
          }
        },
        "required": [
          "category",
          "majorClass",
          "middleClass",
          "minorClass"
        ],
        "additionalProperties": false
      },
      "PartSyncDefault": {
        "type": "string",
        "enum": [
          "同期",
          "非同期",
          "交付版のみ",
          "全体版のみ"
        ]
      },
      "PartMasterReflectDefault": {
        "type": "string",
        "enum": [
          "反映",
          "非反映"
        ]
      },
      "PartHistoryEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "templateId": {
            "type": "string"
          },
          "partKey": {
            "type": "string",
            "description": "版を跨いで安定なパーツ構造キー(文書全体でのアンカー#通し番号)。GrapesJS のコンポーネントid は再採番され不安定なため構造キーで紐づける"
          },
          "user": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "change": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "templateId",
          "partKey",
          "user",
          "timestamp",
          "change"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "editor.sid",
        "description": "ログイン時に発行されるセッション cookie。"
      }
    }
  }
}