# editor/server/src/routes/parts.routes.ts
// =============================================================================
// parts.routes.ts — パーツカタログのルート(エディタ左ペイン)+ パーツ単位履歴
// =============================================================================
import { apiPaths, type PartClassificationQuery } from '@editor/shared';
import { RecordPartChangeRequest } from '@editor/shared/schemas';
import type { FastifyPluginAsync } from 'fastify';
import type { z } from 'zod';
import type { Deps } from '../deps.js';
import { requireAuth, requireEditor } from '../middleware/auth.js';
import { validate } from '../middleware/validate.js';
import * as history from '../repositories/historyRepo.js';
import { actorOf, pickQuery } from './routeHelpers.js';

function toClassQuery(q: Record<string, unknown>): PartClassificationQuery {
  return pickQuery(q, [
    'category',
    'majorClass',
    'middleClass',
    'minorClass',
    'editionType',
  ] as const);
}

type ClassQuery = { Querystring: Record<string, unknown> };
type PartHistoryParams = { Params: { templateId: string } };

export const partsRoutes: FastifyPluginAsync<{ deps: Pick<Deps, 'parts'> }> = async (app, opts) => {
  const { parts } = opts.deps;

  app.get<ClassQuery>(
    apiPaths.partClassificationOptions,
    { preHandler: requireAuth },
    async (request) => {
      return parts.getPartClassificationOptions(toClassQuery(request.query));
    },
  );

  app.get<ClassQuery>(apiPaths.parts, { preHandler: requireAuth }, async (request) => {
    return parts.listParts(toClassQuery(request.query));
  });

  app.get<PartHistoryParams>(apiPaths.partHistory, { preHandler: requireAuth }, async (request) => {
    return history.listPartHistory(request.params.templateId);
  });

  app.post<PartHistoryParams & { Body: z.infer<typeof RecordPartChangeRequest> }>(
    apiPaths.partHistory,
    { preHandler: [requireAuth, requireEditor, validate(RecordPartChangeRequest)] },
    async (request, reply) => {
      const body = request.body;
      await history.recordPartChange(
        request.params.templateId,
        body.partKey,
        body.change,
        actorOf(request),
        body.id,
      );
      return reply.code(204).send();
    },
  );
};